theme: separate out base theme constructs
